Initial Xtext-based setup and grammar definition for Metaflow
This is an Xtext-based implementation of the Metaflow language from
the Culturegraph project. For now, it's basically a port of the
Metaflow ANTLR3 grammar to Xtext, providing an Eclipse-based editor
for Metaflow *.flow files.

To create and run the editor in an Eclipse instance with Xtext 2.3
installed, open the Metatext.xtext grammar file an choose 'Run As'
-> 'Generate Xtext artifacts'. Then, choose the org.lobid.metatext
project, 'Run As' -> 'Eclipse application'.

In the launched workbench, set up a project and copy or create some
Metaflow *.flow files. They will open in the generated editor,
which provides basic syntax coloring, auto completion, and outline.