Generate DOT with Xtend to visualize Flux workflows with Zest
On saving a *.flux file in the editor, the generator will create a
corresponding *.dot file in the project's `src-gen` folder.

With the Zest graph view open and listening to changes in the
workspace (see user guide for usage details), this visualizes the
workflow currently edited in the Flux editor (see also #7).