Show warning, not error for unverifiable, generics-based workflows
See #33. This tweaks one of the test workflows to include a case
where one command outputs `Object`, and the next requires `String`.
Internally, this is based on generics and works, so we display a
warning here, not an error. This commit unifies this new warning
with the previous two (for missing input and output annotations).