Convert `\n` to `<br/>` in Flux command description metadata
The Flux command metadata is used by the auto-suggest popup to show
command details (descriptiopn, input, output). With this change, it
is possible to display a formatted description, e.g. for a list of
supported parameters. For that, the `@Description` of the command's
implementation class needs to contain `\n` characters.